done and will continue next time

+++ server/routers/snippetRouter.js
@@ -1,10 +1,11 @@
 const router= require("express").Router();
 const Snippet= require("../models/snippetModel");
+const auth= require("../middleware/auth");
 
 // get all snippets
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
     try {
-        const snippets= await Snippet.find();
+        const snippets= await Snippet.find({user: req.user});
         res.json(snippets);
     }
     catch(err) {
@@ -13,7 +14,7 @@ router.get("/", async (req, res) => {
 });
 
 // post snippets
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
     try {
         const {title, description, code}= req.body;
 
@@ -21,7 +22,7 @@ router.post("/", async (req, res) => {
             return res.status(400).json({errorMessage: "You need to enter at least a description or code."});
         }
         const newSnippet= new Snippet({
-            title, description, code
+            title, description, code, user: req.user
         });
         
         const savedSnippet= await newSnippet.save();
@@ -34,7 +35,7 @@ router.post("/", async (req, res) => {
 });
 
 // delete snippet by a specific Id
-router.delete("/:id", async (req,res) => {
+router.delete("/:id", auth, async (req,res) => {
     try {
         const snippetId= req.params.id;
         
@@ -46,6 +47,9 @@ router.delete("/:id", async (req,res) => {
         if (!existingSnippet)
             return res.status(400).json({errorMessage: "No snippet with this Id was found. Please contact the developer"})
 
+        if (existingSnippet.user.toString() !== req.user)
+            return res.status(401).json({errorMessage: 'Unauthorized.'});
+
         await existingSnippet.delete();
         
         res.json(existingSnippet);
@@ -56,7 +60,7 @@ router.delete("/:id", async (req,res) => {
 });
 
 // update a snippet by a specific Id
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
     try {
         const {title, description,code}= req.body;
         const snippetId= req.params.id;
@@ -71,6 +75,9 @@ router.put("/:id", async (req, res) => {
         const originalSnippet= await Snippet.findById(snippetId);
         if (!originalSnippet)
             return res.status(400).json({errorMessage: "No snippet with this Id was found. Please contact the developer"});
+
+        if (originalSnippet.user.toString() !== req.user)
+            return res.status(401).json({errorMessage: 'Unauthorized.'});
         
         originalSnippet.title= title;
         originalSnippet.description= description;
